+++ Intern/rayx-core/src/Shader/LineDensity.cpp
@@ -46,8 +46,8 @@ given direction on the grating
 */
 RAYX_FN_ACC
 void RAYX_API RZPLineDensity(Ray r, glm::dvec3 normal, Behaviour::RZP b, double& DX, double& DZ) {
-    int IMAGE_TYPE = int(b.m_imageType);
-    int RZP_TYPE = int(b.m_rzpType);
+    int IMAGE_TYPE = b.m_imageType;
+    int RZP_TYPE = b.m_rzpType;
     double risag = b.m_designSagittalEntranceArmLength;
     double rosag = b.m_designSagittalExitArmLength;
     double rimer = b.m_designMeridionalEntranceArmLength;
@@ -55,7 +55,7 @@ void RAYX_API RZPLineDensity(Ray r, glm::dvec3 normal, Behaviour::RZP b, double&
     double alpha = b.m_designAlphaAngle;
     double beta = b.m_designBetaAngle;
     double WL = 1e-06 * b.m_designWavelength;  // source energy/design energy
-    double Ord = b.m_orderOfDiffraction;
+    int Ord = b.m_orderOfDiffraction;
 
     double FX = normal.x;
     double FY = normal.y;
